+++ docs/REQUIREMENTS.md
@@ -28,10 +28,10 @@
     - [x] 카드번호 뒷 8자리는 온점으로 표시한다.
     - [x] "?" 버튼에 hover 상태면 cvc에 대한 설명을 보여줌.
     - [x] (예외) 입력값이 잘못 들어오면 오류 메시지를 띄워준다.
-    - [ ] 카드를 클릭하면 bottom sheet을 연다.
-    - [ ] bottom sheet에서는 8개의 은행을 사용자가 선택할 수 있다. 은행을 선택하면 카드 좌측 상단에 선택된 카드 회사 이름이 뜨고, 카드 회사의 메인 색깔로 변경된다.
-    - [ ] 카드 등록이 완료되면 카드에 별칭을 부여할 수 있다.
-      - [ ] 별칭은 등록할 수도 있고 안할수도 있다.
+    - [x] 카드를 클릭하면 bottom sheet을 연다.
+    - [x] bottom sheet에서는 8개의 은행을 사용자가 선택할 수 있다. 은행을 선택하면 카드 좌측 상단에 선택된 카드 회사 이름이 뜨고, 카드 회사의 메인 색깔로 변경된다.
+    - [x] 카드 등록이 완료되면 카드에 별칭을 부여할 수 있다.
+      - [x] 별칭은 등록할 수도 있고 안할수도 있다.
 
 ### 카드 컴포넌트
