+++ src/stories/AddCardNumberInput.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta } from '@storybook/react';
 
 import CardNumberInput from '../pages/AddCard/components/CardNumberInput';
-import { APP_WIDTH } from './constants';
+import { APP_WIDTH } from '../utils/constants';
 import { isValidCardNumber } from '../pages/AddCard/domain/dispatcher';
 import useInput from '../hooks/useInput';
