--- src/stories/constants.ts
@@ -1,1 +1,0 @@
-export const APP_WIDTH = '318px';
